+++ src/stream/src/common/infallible_expr.rs
